* Person : Force upscale ratio 1:1 기능 추가 * 테스트용 숨겨진 기능 2가지 추가 * Setting : Allways use FULL, VAE type for encode when detail anything. (v1.6.0)

+++ sd_bmab/process.py
@@ -623,5 +623,5 @@ def process_upscale_inner(image, s, p, args):
 
 	w = image.width
 	h = image.height
-	img = images.resize_image(0, image, w * ratio, h * ratio, upscaler)
+	img = images.resize_image(0, image, int(w * ratio), int(h * ratio), upscaler)
 	return img.convert('RGB')
